fixing some small issues

+++ calculateGini.cpp
@@ -6,12 +6,12 @@
 using namespace std;
 
 float calcSurvProp(const std::vector<Passenger>& dataSet) {
-    int survived = 0;
+    float survived = 0;
 
     for (int i = 0; i < dataSet.size(); ++i) {
         if (dataSet[i].Surv) survived++;
     }
-    return static_cast<float>(survived) / static_cast<float>(dataSet.size());
+    return survived / static_cast<float>(dataSet.size());
 }
 
 float calcBinaryGini(float P){
